refactor: 각 화면 별 패키지 분리
