+++ app/src/main/java/woowacourse/movie/activities/movielist/adapter/MovieListAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.adapter
+package woowacourse.movie.activities.movielist.adapter
 
 import android.view.LayoutInflater
 import android.view.View
