--- app/src/main/java/woowacourse/movie/view/adapter/OnBookClickListener.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.view.adapter
-
-import woowacourse.movie.model.Movie
-
-interface OnBookClickListener {
-    fun onClick(item: Movie)
-}
