+++ app/src/main/res/layout-land/activity_ticketing_result.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".view.activities.TicketingResultActivity">
+    tools:context=".view.activities.ticketingresult.TicketingResultActivity">
 
     <View
         android:id="@+id/ticketing_header_container"
