+++ app/src/main/res/layout/activity_movie_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activities.MovieListActivity">
+    tools:context=".activities.movielist.MovieListActivity">
 
     <ListView
         android:id="@+id/lv_movies"
