+++ app/src/main/res/layout/activity_ticketing.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activities.TicketingActivity">
+    tools:context=".activities.ticketing.TicketingActivity">
 
     <ScrollView
         android:id="@+id/sv_movie_info"
